tracing_tree

Struct HierarchicalLayer

pub struct HierarchicalLayer<W = fn() -> Stderr, FT = ()>
where
    W: for<'writer> MakeWriter<'writer> + 'static,
    FT: FormatTime,
{ /* private fields */ }

Implementations

impl HierarchicalLayer<fn() -> Stderr>

pub fn new(indent_amount: usize) -> Self

impl<W, FT> HierarchicalLayer<W, FT>
where W: for<'writer> MakeWriter<'writer> + 'static, FT: FormatTime,

pub fn with_ansi(self, ansi: bool) -> Self

Enables terminal colors, boldness and italics.

pub fn with_writer<W2>(self, make_writer: W2) -> HierarchicalLayer<W2, FT>
where W2: for<'writer> MakeWriter<'writer>,

pub fn with_indent_amount(self, indent_amount: usize) -> Self

pub fn with_indent_lines(self, indent_lines: bool) -> Self

Renders an ascii art tree instead of just using whitespace indentation.

pub fn with_timer<FT2: FormatTime>( self, timer: FT2, ) -> HierarchicalLayer<W, FT2>

Specifies how to measure and format time at which event has occurred.

pub fn with_targets(self, targets: bool) -> Self

Whether to render the event and span targets. Usually targets are the module path to the event/span macro invocation.

pub fn with_thread_ids(self, thread_ids: bool) -> Self

Whether to render the thread id in the beginning of every line. This is helpful to untangle the tracing statements emitted by each thread.

pub fn with_thread_names(self, thread_names: bool) -> Self

Whether to render the thread name in the beginning of every line. Not all threads have names, but if they do, this may be more helpful than the generic thread ids.

pub fn with_wraparound(self, wraparound: usize) -> Self

Resets the indentation to zero after wraparound indentation levels. This is helpful if you expect very deeply nested spans as otherwise the indentation just runs out of your screen.

pub fn with_verbose_entry(self, verbose_entry: bool) -> Self

Whether to print the currently active span’s message again before entering a new span. This helps if the entry to the current span was quite a while back (and with scrolling upwards in logs).

pub fn with_verbose_exit(self, verbose_exit: bool) -> Self

Whether to print the currently active span’s message again before dropping it. This helps if the entry to the current span was quite a while back (and with scrolling upwards in logs).

pub fn with_span_retrace(self, enabled: bool) -> Self

Whether to print the currently active span’s message again if another span was entered in the meantime This helps during concurrent or multi-threaded events where threads are entered, but not necessarily exited before other divergent spans are entered and generating events.

pub fn with_deferred_spans(self, enabled: bool) -> Self

Defers printing span opening until an event is generated within the span.

Avoids printing empty spans with no generated events.

pub fn with_span_modes(self, enabled: bool) -> Self

Prefixes each branch with the event mode, such as open, or close

pub fn with_bracketed_fields(self, bracketed_fields: bool) -> Self

Whether to print {} around the fields when printing a span. This can help visually distinguish fields from the rest of the message.